chore: split legacy config test from default test task
Move grpc_server compatibility coverage behind a legacy build tag and add
`task test:legacy`, so `task test` no longer depends on legacy paths.
