+++ docs/legacy-removal.md
@@ -72,5 +72,5 @@ gateway_server:
 - [x] `internal/configs/components/grpc_server.yaml` 已移除，统一 `gateway_server.yaml`
 - [x] `lavabuilder grpc` 通过 `gatewayserver.LoadConfig` 加载 YAML 并对 `grpc_server` 打废弃警告
 - [x] 架构文档仅描述 `gateway_server`
-- [ ] `task test` 不依赖 legacy 路径（或单独 `task test:legacy`）
+- [x] `task test` 不依赖 legacy 路径；兼容用例移至 `task test:legacy`
 - [x] 删除 `grpcs` 包（v2，不再等待 v3）
